+++ startup/index.js
@@ -1,5 +1,6 @@
-const server = require ('./server');
-
+const app = require ('./server');
+const {WebSocketServer} = require ('ws');
+const wsserver = app.websocket.wsserver
 try{
     run();
 } catch (error){
@@ -10,6 +11,12 @@ finally{
 }
 
 async function run (){
-    server.listen(4000);
+   const server = app.server.listen(4000);
+
+    server.on('upgrade', (request, socket, head) => {
+        wsserver.handleUpgrade(request, socket, head, function done(ws) {
+            wsserver.emit('connection', ws, request);
+        });
+    });
     console.log(`Listening on port 4000`);
 }
